+++ src/blocks/search-filter/save.js
@@ -1,0 +1,5 @@
+function save() {
+	return null;
+}
+
+export default save;